data_collector/requirements.txt
ccxt
polars
python-dotenv
sqlalchemy
psycopg2-binary